requirements-web.txt
flask==2.3.2
waitress==2.1.2
python-jose==3.3.0
python-jwt==4.0.0
websocket-client==1.6.1
